DLS-10769 Abstract away common elements from tests
